requirements-dev-qt.txt
pytest-qt
pytest-xvfb
